build: improve FetchContent usage, avoid VVL build
We don't need the entire Vulkan-ValidationLayers build (which has
a lot of nested dependencies that can drive source path names to be
longer than Windows allows, without long path settings).  Using
FetchContent_MakeAvailable() causes the project and all its dependencies
to be added to the project build.  Switching to FetchContent_Populate()
allows the needed source files to be available without triggering the
deep dependency builds.

But FetchContent documents that FetchContent_MakeAvailable() and
FetchContent_Populate() set variables <lowercaseName>_POPULATED and
<lowercaseName>_SOURCE_DIR.  We had been using non-lower-case
variable names, e.g. Vulkan-ValidationLayers_SOURCE_DIR, which
are undocumented, but happen to be set in MakeAvailable() but not
in Populate().  These changes drop all the FetchContent project
names to lowercase, to both comply with the FetchContent documentation
and for consistency.

Also, FetchContent documentation encourages projects to declare all
content before fetching any of it, to ensure that the project's
declarations properly override any declarations in fetched projects.

Finally, attempting to compile with VS2015 gets past the long
path issues, but still failed on std::min:

    c:\j\workspace\vulkanprofiles\64-bit\release\vulkanprofiles\library\test\vulkan_profiles.hpp(880): error C2039: 'min': is not a member of 'std' [C:\j\workspace\VulkanProfiles\64-bit\Release\VulkanProfiles\build\library\test\library_test_profile_api.vcxproj]

The web identified that

    #include <algorithm>

was needed to allow the build to succeed.

This doesn't fix the SDK "flat build" dependency, which will be addressed
separately.