tests: add environment variables for devsim test cases
devsim tests need to be able to set the VK_LAYER_PATH and other
OS-dependent environment variables in order to find and use the project's
devsim layer.  We use CMake's set_tests_properties() to set the environment.
This is sort of clunky and very Windows-centric, but it can be made to
work on Linux.

Also, there were multiple DevsimTest() definitions, which was confusing
at best, and makes it difficult to know which DevsimTest was actually
being used.  I removed the one that currently is not being referenced.
If we do need DevsimTest() in mutliple subdirectories, it should be
moved to the top-level CMakeLists.txt at some future time.

- removed DevsimTest() function from device_simulation/tests/CMakeLists.txt
  to ensure that only one copy exists
- add temporary test target to spit out devsim test environments
- add OS-dependent test environment setting