android: get the Android build working
With these changes, an Android build on a working Android machine can be made
with:

    $ cd VulkanProfiles
    $ mkdir android-build
    $ cd android-build
    $ cmake .. -DCMAKE_TOOLCHAIN_FILE=$ANDROID_NDK_HOME/build/cmake/android.toolchain.cmake -DANDROID_PLATFORM=28
    $ cmake --build .

The android.toolchain.cmake has almost all the boilerplate required to build
Android in a standard CMake process.  We need to specify ANDROID_PLATFORM because
the default platform that supports Vulkan (26) doesn't support VkGetPhysicalDeviceFeatures2,
VkGetPhysicalDeviceProperties2, and VkGetPhysicalDeviceFormatProperties2, which are
needed by the project.  Android platform 28 is the first platform to support these.

Android builds should not depend on Vulkan-Loader, because we don't compile
a separate loader on Android - we always use the system loader.  We also have
no need to build the library_generator on Android - this binary is used to
create the vulkan_profiles.h, vulkan_profiles.hpp, and vulkan_profiles.cpp
files, which should be done on the host machine, not an Android device.

We have to turn off Vulkan-Loader when compiling for Android, since
it will try to build X11 and XCB support on a Linux host, neither of
which works for Android.

Since those generated files already exist as checked-in source, there's
actually no need for *any* platform to build library_generator, unless
you're regenerating or ensuring that the current generators produce the
same output as the checked-in source.  For now, though, I'm just not
building the generator on Android - we may clean it up more later.

Android layers need to have the "-llog" library, to get the
__android_print_log() function.  This was added to the add_vk_layer macro.
There were two copies of the add_vk_layer macro, so the macro was promoted
into the parent CMakeLists.txt.

Note that these changes allow VulkanProfiles code to be built with the Android
compiler, but additional changes will be required to create functional Android
packages that can actually be tested.  Also note that the VulkanTests harness
will have to be updated to use this Android build process.

CMakeLists.txt:
- Android builds do not fetch Vulkan-Loader
- promote add_vk_layer macro to here instead of having multiple copies
- fix add_vk_layer macro to add -llog on Android

device_simulation/CMakeLists.txt:
- remove add_vk_layer macro

layer/CMakeLists.txt:
- remove add_vk_layer macro

library/CMakeLists.txt
- don't build the generator on Android

library/test/CMakeLists.txt:
- use common set of test_libraries
- add -llog -landroid -ldl on Android