build: always use at least c++11
Recent changes require C++ 2011 or better.  This isn't an issue with
more recent compilers, but old compilers (e.g. Ubuntu 16.04, which is
currently used for building the SDK) need the -std=c++11 switch in order
to compile correctly.

This switch was set explicitly in three lower-level CMakeLists.txt, but
really needs to be in the top-level CMakeLists.txt in order to affect
the whole project.

Also, instead of setting CMAKE_CXX_FLAGS, it's preferred to set
CMAKE_CXX_STANDARD and let CMake decide the appropriate platform-specific
flags to apply.