build: even flatter SDK build
The former attempt to flatten the SDK build (by flattening only the
Vulkan dependencies) works for the Windows SDK build, but is insufficient
for the Linux SDK build.

Linux package builds require that clean source packages be available for
all binary packages, and cannot have source we do not own in the source
tree.  Also, the Linux container build is hardened, and cannot access
the network to download after the initial source checkout.

The json dependency isn't actually being used, so it has been removed.

The googletest dependency is only needed when building tests, so
it has been hidden behind a new PROFILES_BUILD_TESTS configuration
variable (so named to avoid conflict with the existing
Vulkan-ValidationLayers BUILD_TESTS variable), which can be
turned off for SDK builds.

The jsoncpp dependency is interesting; the build artifacts used are
actually source generated by this project build, not anything that's
in the jsoncpp source tree.  Putting it anywhere in a build tree would
violate the Linux requirements.  So we're treating it as generated source,
and checking the generated code directly into the repository.  The default
behavior remains the same, but if the new CMake switch REGENERATE_JSONCPP=OFF
is passed, the checked-in code will be used and no attempt will be made to
fetch or generate jsoncpp content.  (Given that the last change to jsoncpp
was made in 2010, this is probably not going to become an update issue.)

CMakeLists.txt:
- remove json dependency
- add PROFILES_BUILD_TESTS configuration option
- only fetch googletest if PROFILES_BUILD_TESTS is ON
- only add test subdirectories if PROFILES_BUILD_TESTS is ON
- Only fetch jsoncpp content if option is set
- Only regenerate jsoncpp.cpp, json.h, and json-forwards.h if option is set

layer/CMakeLists.txt:
- only add test subdirectories if PROFILES_BUILD_TESTS is ON
- reference generated jsoncpp source and header files

layer/tests/CMakeLists.txt:
- reference generated jsoncpp source and header files

library/CMakeLists.txt:
- only add test subdirectories if PROFILES_BUILD_TESTS is ON

profiles/test/CMakeLists.txt:
- reference generated jsoncpp source and header files

generated-jsoncpp/jsoncpp.cpp, generated-jsoncpp/json/*.h:
- generated jsoncpp source and header files