generation: fix python reference
The genvp.py script was being called with "python".  This is incorrect
on Linux builds, as "python" is typically Python 2.

The correct usage is ${PYTHON_EXECUTABLE}, as set by the call to:
    find_package(PythonInterp 3 REQUIRED)

Also, genvp.py uses typing.OrderedDict, which is not available until
Python3.7.2.  So I've updated the Python interpreter requirement.