issue 167: add REGENERATE_LAYER and FATAL_WARNINGS switches
Addressing issue:

    Build of Vulkan-Profiles fails on Fedora 34
    https://github.com/KhronosGroup/Vulkan-Profiles/issues/167

There are two issues here:

-   Building the Vulkan-Profiles layer requires the Python jsonschema module.

-   Fedora compilers give more warnings that Ubuntu compilers.

The build works if the Python "jsonschema" module is added to the
building system, and if -Werror is removed from layers/CMakeLists.txt.

To avoid requiring users to install "jsonschema", a new switch
REGENERATE_LAYER is provided.  If ON (the default), the layer code
for profiles.cpp is generated normally via the gen_layer.py script.
If OFF, the committed repository code for profiles.cpp is used directly
without first regenerating.  The "vulkansdk" script and the SDK
builds will explicitly turn this OFF.

To handle differing warnings between compilers, a new switch
FATAL_WARNINGS is provided; if ON (the default) all compile warnings
in the project will become errors.  If OFF, compile warnings are
issued but will not block the build.  The "vulkansdk" script will
explicitly turns this OFF.

These changes include additional fixes for existing code that provoked
compiler warnings:

- layer-utils/vk_layer_settings.cpp: add cast to avoid:
     warning C4244: '=': conversion from 'int' to 'char', possible loss of data

and a lot of Windows compiler instructions to ignore particular warnings.
(These warnings should be fixed as a later effort.)