layer: Combine LogMessage and format()
Instead of having each caller of LogMessage perform formatting, simply
pass the arguments to LogMessage and let it do the necessary formatting.
This is a large performance boost because formatting will now only
occur if the message is to be written out. Before, many messages would
format their output then call LogMessage, which would exit early due
to the current debug level.