layer: Fixup PhysicalDeviceData
Emplacing into a map requires that the constructor for the object be
public. Thus, the constructor for PhysicalDeviceData had to be made
public in order for the emplace() to not have to move/copy
PhysicalDeviceData.

This commit also marks both the copy constructor and copy-assign
operator as deleted while also making them public, since there is no
reason for deleted functions to be private.