layer: support loading profiles with capability-alternatives
The "capabilities" property is a list of either strings or lists of
strings. When it's a list of strings, it seems to be intended to mean
multiple alternatives that can fulfil the same requirement.

In order to pass this information on to ReadProfile, we need to turn it
into a list of lists of strings only. When there's just one alternative,
we use a list of length one. The JSON syntax is really just a short-form
here.

So let's parse this correctly, and loop over the alternatives and try
them in order instead of just failing to parse the profile.

This was noticed when trying to simulate the various profiles listed in
the zink requirements:

https://gitlab.freedesktop.org/mesa/mesa/-/blob/main/src/gallium/drivers/zink/VP_ZINK_requirements.json