Publish new Android AVP 2025 and VRA17 profiles
AVP 2025 reflects combined Vulkan functionality that is available on
at least 80% of active Android devices in the latter part of 2025.

VRA17 is a collection of functionality that is mandated for chipsets
that launch (or renew Google Requirements Freeze) on Android 17.  It
reflects the needs of some Android software developers, that GPU
vendors can deliver in time for Android 17.